+++ Cmf.Custom.Help/src/packages/cmf.docs.area.amsosram/assets/releasenotes.md
@@ -18,5 +18,7 @@ List of delivered packages:
 * [AMSOsram-20220314-1.2.0](/AMSOsram/releasenotes>AMSOsram-20220314-1.2.0)
 * [AMSOsram-20220328-1.3.0](/AMSOsram/releasenotes>AMSOsram-20220328-1.3.0)
 * [AMSOsram-20220411-1.4.0](/AMSOsram/releasenotes>AMSOsram-20220411-1.4.0)
+* [AMSOsram-20220426-1.5.0](/AMSOsram/releasenotes>AMSOsram-20220426-1.5.0)
+* [AMSOsram-20220509-1.6.0](/AMSOsram/releasenotes>AMSOsram-20220509-1.6.0)